+++ cobalt/android/apk/app/src/main/java/dev/cobalt/coat/CommandLineOverrideHelper.java
@@ -98,6 +98,10 @@ public final class CommandLineOverrideHelper {
         // It is important to use a feature override instead of the
         // rendering switch, to make sure certain devices are excluded.
         paramOverrides.add("DefaultPassthroughCommandDecoder");
+        // Compositor switch to reduce prepaint tile cache size
+        paramOverrides.add("SmallerInterestArea");
+        paramOverrides.add("ReclaimPrepaintTilesWhenIdle");
+        paramOverrides.add("ReclaimOldPrepaintTiles");
 
         return paramOverrides;
     }