+++ cobalt/browser/h5vcc_updater/h5vcc_updater_impl.cc
@@ -1,0 +1,126 @@
+// Copyright 2025 The Cobalt Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "cobalt/browser/h5vcc_updater/h5vcc_updater_impl.h"
+
+#include "base/functional/callback.h"
+#include "build/build_config.h"
+
+#if BUILDFLAG(USE_EVERGREEN)
+#include "cobalt/updater/updater_module.h"
+#endif
+
+#if BUILDFLAG(IS_STARBOARD)
+#include "cobalt/configuration/configuration.h"
+#include "starboard/common/system_property.h"
+#include "starboard/system.h"
+#endif
+
+namespace h5vcc_updater {
+
+H5vccUpdaterImpl::H5vccUpdaterImpl(
+    content::RenderFrameHost& render_frame_host,
+    mojo::PendingReceiver<mojom::H5vccUpdater> receiver)
+    : content::DocumentService<mojom::H5vccUpdater>(render_frame_host,
+                                                    std::move(receiver)) {
+  DETACH_FROM_THREAD(thread_checker_);
+}
+
+H5vccUpdaterImpl::~H5vccUpdaterImpl() {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+}
+
+void H5vccUpdaterImpl::Create(
+    content::RenderFrameHost* render_frame_host,
+    mojo::PendingReceiver<mojom::H5vccUpdater> receiver) {
+  new H5vccUpdaterImpl(*render_frame_host, std::move(receiver));
+}
+
+void H5vccUpdaterImpl::GetUpdaterChannel(GetUpdaterChannelCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::SetUpdaterChannel(const std::string& channel,
+                                         SetUpdaterChannelCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetUpdateStatus(GetUpdateStatusCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::ResetInstallations(ResetInstallationsCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetInstallationIndex(
+    GetInstallationIndexCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetAllowSelfSignedPackages(
+    GetAllowSelfSignedPackagesCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::SetAllowSelfSignedPackages(
+    bool allow_self_signed_packages,
+    SetAllowSelfSignedPackagesCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetUpdateServerUrl(GetUpdateServerUrlCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+#if BUILDFLAG(USE_EVERGREEN)
+  std::string url =
+      cobalt::updater::UpdaterModule::GetInstance()->GetUpdateServerUrl();
+  std::move(callback).Run(url);
+#else
+  NOTIMPLEMENTED();
+#endif
+}
+
+void H5vccUpdaterImpl::SetUpdateServerUrl(const std::string& update_server_url,
+                                          SetUpdateServerUrlCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetRequireNetworkEncryption(
+    GetRequireNetworkEncryptionCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::SetRequireNetworkEncryption(
+    bool require_network_encryption,
+    SetRequireNetworkEncryptionCallback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+void H5vccUpdaterImpl::GetLibrarySha256(unsigned short index,
+                                        GetLibrarySha256Callback callback) {
+  DCHECK_CALLED_ON_VALID_THREAD(thread_checker_);
+  NOTIMPLEMENTED();
+}
+
+}  // namespace h5vcc_updater