+++ starboard/android/shared/media_codec_bridge.cc
@@ -181,18 +181,17 @@ MediaCodecBridge::CreateVideoMediaCodecBridge(
     jobject j_surface,
     jobject j_media_crypto,
     const SbMediaColorMetadata* color_metadata,
+    bool enable_frame_renderer_listener,
     bool require_secured_decoder,
     bool require_software_codec,
     std::optional<int> tunnel_mode_audio_session_id,
     bool force_big_endian_hdr_metadata,
     int max_video_input_size,
-    bool enable_output_checker,
     bool skip_video_frames_over_60_fps) {
   if (max_frame_size) {
     SB_CHECK_GT(max_frame_size->width, 0);
     SB_CHECK_GT(max_frame_size->height, 0);
   }
-
   const char* mime = SupportedVideoCodecToMimeType(video_codec);
   if (!mime) {
     return Failure(std::string("Unsupported mime for codec: ") +
@@ -208,30 +207,18 @@ MediaCodecBridge::CreateVideoMediaCodecBridge(
   std::string decoder_name =
       MediaCapabilitiesCache::GetInstance()->FindVideoDecoder(
           mime, must_support_secure, must_support_hdr, require_software_codec,
-          must_support_tunnel_mode,
-          /* frame_width = */ 0,
-          /* frame_height = */ 0,
-          /* bitrate = */ 0,
-          /* fps = */ 0);
+          must_support_tunnel_mode);
   if (decoder_name.empty() && color_metadata) {
     // On second pass, forget HDR.
     decoder_name = MediaCapabilitiesCache::GetInstance()->FindVideoDecoder(
-        mime, must_support_secure, /* must_support_hdr = */ false,
-        require_software_codec, must_support_tunnel_mode,
-        /* frame_width = */ 0,
-        /* frame_height = */ 0,
-        /* bitrate = */ 0,
-        /* fps = */ 0);
+        mime, must_support_secure, /*must_support_hdr=*/false,
+        require_software_codec, must_support_tunnel_mode);
   }
   if (decoder_name.empty() && require_software_codec) {
     // On third pass, forget software codec required.
     decoder_name = MediaCapabilitiesCache::GetInstance()->FindVideoDecoder(
-        mime, must_support_secure, /* must_support_hdr = */ false,
-        /* require_software_codec = */ false, must_support_tunnel_mode,
-        /* frame_width = */ 0,
-        /* frame_height = */ 0,
-        /* bitrate = */ 0,
-        /* fps = */ 0);
+        mime, must_support_secure, /*must_support_hdr=*/false,
+        /*require_software_codec=*/false, must_support_tunnel_mode);
   }
 
   if (decoder_name.empty()) {
@@ -291,7 +278,7 @@ MediaCodecBridge::CreateVideoMediaCodecBridge(
       max_frame_size ? max_frame_size->height : -1, j_surface_local,
       j_media_crypto_local, j_color_info,
       tunnel_mode_audio_session_id.value_or(TUNNEL_MODE_AUDIO_SESSION_ID_NONE),
-      max_video_input_size, enable_output_checker,
+      max_video_input_size, enable_frame_renderer_listener,
       skip_video_frames_over_60_fps, j_create_media_codec_bridge_result);
 
   ScopedJavaLocalRef<jobject> j_media_codec_bridge(
@@ -503,12 +490,6 @@ void MediaCodecBridge::OnMediaCodecFrameRendered(
 
 void MediaCodecBridge::OnMediaCodecFirstTunnelFrameReady(JNIEnv* env) {
   handler_->OnMediaCodecFirstTunnelFrameReady();
-}
-
-// static
-jboolean MediaCodecBridge::IsFrameRenderedCallbackEnabled() {
-  JNIEnv* env = AttachCurrentThread();
-  return Java_MediaCodecBridge_isFrameRenderedCallbackEnabled(env);
 }
 
 MediaCodecBridge::MediaCodecBridge(Handler* handler) : handler_(handler) {