+++ starboard/shared/x11/application_x11.cc
@@ -779,9 +779,7 @@ void ApplicationX11::Composite() {
           }
           current_video_frames_[player] = cpu_video_frame;
         }
-        window->CompositeVideoFrame(frame_info.x, frame_info.y,
-                                    frame_info.width, frame_info.height,
-                                    cpu_video_frame);
+        window->CompositeVideoFrame(frame_info.rect, cpu_video_frame);
       }
       window->EndComposite();
     }
@@ -793,10 +791,7 @@ void ApplicationX11::Composite() {
 void ApplicationX11::AcceptFrame(SbPlayer player,
                                  const scoped_refptr<VideoFrame>& frame,
                                  int z_index,
-                                 int x,
-                                 int y,
-                                 int width,
-                                 int height) {
+                                 const Rect& rect) {
   std::lock_guard lock(frame_mutex_);
 
   if (frame->is_end_of_stream()) {
@@ -813,19 +808,13 @@ void ApplicationX11::AcceptFrame(SbPlayer player,
 
 void ApplicationX11::PlayerSetBounds(SbPlayer player,
                                      int z_index,
-                                     int x,
-                                     int y,
-                                     int width,
-                                     int height) {
+                                     const Rect& rect) {
   std::lock_guard lock(frame_mutex_);
 
   FrameInfo& frame_info = next_video_bounds_[player];
   frame_info.player = player;
   frame_info.z_index = z_index;
-  frame_info.x = x;
-  frame_info.y = y;
-  frame_info.width = width;
-  frame_info.height = height;
+  frame_info.rect = rect;
 
   // The bounds should only take effect once the UI frame is submitted. But we
   // also apply the bounds immediately so that there is no flicker.